Flesh out some of the Dojo page and export it
Not sure yet whether we will want to use this or not, but putting it
up just in case.

I think I've changed some of the export settings for the index
page (so eg. the publish date at the bottom disappears). We should
make a proper publish project to configure this stuff.