Ensure the htmlize syntax highlighting is used on export
This provides syntax highlighting for the dojo code snippets.
Right now its not deterministic - it depends on the current theme
used. I've just exported it initially with my custom gruvbox theme - I
think it will be fine for now for people to just export it with
whatever theme they have.